+++ changeLog.md
@@ -1,3 +1,5 @@
 ## changeLog
 
+- backend depoly + insert stock_main 
 - main target (260611)
+